Separate ManageAlerts into components and add alert selection prop logic
